Changed method parameter types.

+++ src/de/mrapp/android/view/util/Condition.java
@@ -56,23 +56,22 @@ public final class Condition {
 	}
 
 	/**
-	 * Ensures, that an {@link Integer} value is at least a specific value.
+	 * Ensures, that a {@link Long} value is at least a specific value.
 	 * Otherwise an {@link IllegalArgumentException} with a specific message
 	 * will be thrown.
 	 * 
 	 * @param value
-	 *            The value, which should be checked, as an {@link Integer}
-	 *            value
+	 *            The value, which should be checked, as a {@link Long} value
 	 * @param referenceValue
-	 *            The value, the given value must be at least, as an
-	 *            {@link Integer} value
+	 *            The value, the given value must be at least, as a {@link Long}
+	 *            value
 	 * @param exceptionMessage
 	 *            The message of the {@link IllegalArgumentException}, which is
 	 *            thrown, if the given value is less than the reference value,
 	 *            as a {@link String}
 	 */
-	public static void ensureAtLeast(final int value, final int referenceValue,
-			final String exceptionMessage) {
+	public static void ensureAtLeast(final long value,
+			final long referenceValue, final String exceptionMessage) {
 		if (value < referenceValue) {
 			throw new IllegalArgumentException(exceptionMessage);
 		}
